Add random quote machine repository
The HTML has a title, div for the quotes to live in and the two
buttons; one to change the quote and the other for share it on twitter.
Also I tried to work with HTML5 tags; header, section and footer. I
have started to add very few styles mainly to the body, still not sure
about the styling yet. Also I have added the script file which is not
fully working yet, in console log the random quotes are working but
still to link that to the DOM with the button.