Add javascript correct syntax first version
I have fixed js file. I have selected the quotesWrapper id and then I
used innerHTML property and assigned another function with one
parameter to it. I passed an argument of randomQuotes to it. Then my
second function which is quotesToString with parameter quotes has
markups and access to the objects’ properties with dot notation. Then
at the end I have selected the location where the quotes are going to
be displayed on the page; quotesWrapper id and then addEventListener to
it with type click and listener which is getRandomQuote function.