Add more quotes and fix click button
I have added more quotes to the quotes array and also I solved a
problem which is when I click on the quotes wrapper element the quote
changes and not just changed when I click on the button. I figured out
that I need to replace the wrapperEl with button itself. Now is working
perfectly.