Resize text and change buttons position
I have changed buttons position to be fixed top and not bottom like I
used to have them. I added a new font -family for the quotes but I am
still not sure about the text size yet for mobile first design. Plus I
need to style the author so make the margin between the quote and
author less and set it to a different font-family.