Continue adding js solution version two
I have worked on the script with Oliver and it was the first time to
read about Fisher-Yates shuffle. So the app shuffles the quotes from
the beginning then when it is empty, it clones the sourceQuotes and
call the shuffle function to  restart the cycle again and again. Using
shift method to remove the first element from the currentQuotes array
which is already will be displayed when we reset the quotes so it is
not going to be added in the getQuote function cycle. Also solved the
tweet problem by targeting the href for the anchor (in html)  and
assign to it the hyperlink which will have status equals the status I
created, mainly a dynamic const status that has the currentQuote. So
just accessed key/value object in a template literal and used
interpolation and a dash for the author. I HAVE LEARNED MANY THINGS
TODAY!