Update styling for the app
I have removed all the @import and added the links in html for fonts
based on a tutorial I have watched about performance optimisation. Also
I have changed the colour of the buttons and when focus plus the style
of author to italic.  Next is doing the last styling for the
quotes-wrapper when click the new quote button.  Then the media queries
and done the app.